An initial style guide configuration using React
This is a stretch for me right now, but we'll see how it goes.

Webpack is used to apply Babel to the JSX and ES2015 JavaScript
code so that it can be built into a final script that browsers
can actually deal with.

This seems a little strange to me, but I like the idea of writing
modern JS and having it compile backward.

We'll still want to hook-up a Grunt task that builds based on the
webpack config any time a JS or JSX file is updated.